script/requirements.txt
transformers
protobuf
sentencepiece
chromadb
sentence-transformers
python-dotenv